+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## []
 ### Changed
-- Serial interface output status can now be configured
+- Serial interface output can now be configured to be on/off
 - Added sample rate to the Relay Controller (compute every 100ms)
 - Rename configuration SPI_DO_PIN to SPI_CS_PIN
 
 ### Added
 - Added new method to the BaseDisplay interface to explicitly print the buffer to the display
+- BaseDisplay interface allow definition of an optional custom print function
 
 ### Fixed
 - Serial baudrate configuration is now used to initialise the serial interface